refactor: apply thunks and api renames
