+++ src/views/Main/Main.tsx
@@ -6,7 +6,7 @@ import BoardPreview from '../../components/BoardPreview/BoardPreview';
 import BoardCreate from '../../components/BoardCreate/BoardCreate';
 import useTypedSelector from '../../hooks/useTypedSelector';
 import useAppDispatch from '../../hooks/useAppDispatch';
-import { getBoardsList } from '../../redux/thunks/thunks';
+import { getBoardsList } from '../../redux/thunks/boardThunks';
 
 const Main = () => {
   const navigate = useNavigate();
